Практическая №9

+++ pz9/Car.java
@@ -3,13 +3,13 @@ package pz9;
 public abstract class Car {
 
     private String type;
-    private int rate;
+    private double rate;
 
-    public Car(String type, int rate){
+    public Car(String type, double rate){
         this.type = type;
         this.rate = rate;
     }
 
     public String getType(){return type;}
-    public int getRate(){return rate;}
+    public double getRate(){return rate;}
 }
